:mag: :hammer: fix e2e test for nerdctl refactor
